requirements.txt
requests==2.26.0
PyGithub==1.55
openai==1.51.0